refactor : Car 기능에 대해 리팩토링한다
