+++ src/main/java/racingcar/README.md
@@ -1,6 +1,6 @@
 # 구현 명세
 - [ ] 사용자에게서 몇 대의 자동차로 몇 번의 이동을 할 것인지 입력받는다
 - [ ] 자동차가 움직인다
-  - [ ] random 값을 구해 4이상일 경우 전진한다
+  - [X] random 값을 구해 4이상일 경우 전진한다
   - [ ] 자동차들의 위치를 출력한다
   - [ ] 입력한 횟수만큼 반복한다
