+++ src/main/java/racingcar/application/Accelerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 public interface Accelerator {
   int accelerate();
