+++ src/main/java/racingcar/application/RandomAccelerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 import java.util.Random;
